+++ rpxy-lib/src/message_handler/handler_manipulate_messages.rs
@@ -1,4 +1,4 @@
-use super::{HttpMessageHandler, handler_main::HandlerContext, utils_headers::*, utils_request::update_request_line};
+use super::{HttpMessageHandler, handler_main::HandlerContext, header_ops::*, request_ops::update_request_line};
 use crate::{
   backend::{BackendApp, UpstreamCandidates},
   constants::RESPONSE_HEADER_SERVER,
@@ -82,13 +82,21 @@ where
     };
 
     let original_uri = req.uri().clone();
+    let original_host_header = req.headers().get(header::HOST).cloned();
     let headers = req.headers_mut();
     // delete headers specified in header.connection
     remove_connection_header(headers);
     // delete hop headers including header.connection
     remove_hop_header(headers);
     // X-Forwarded-For (and Forwarded if exists)
-    add_forwarding_header(headers, client_addr, listen_addr, tls_enabled, &original_uri)?;
+    add_forwarding_header(
+      headers,
+      client_addr,
+      listen_addr,
+      tls_enabled,
+      &original_uri,
+      &self.globals.proxy_config.trusted_forwarded_proxies,
+    )?;
 
     // Add te: trailer if te_trailer
     if contains_te_trailers {
@@ -96,7 +104,7 @@ where
     }
 
     // by default, add "host" header of original server_name if not exist
-    if req.headers().get(header::HOST).is_none() {
+    if original_host_header.is_none() {
       let org_host = req.uri().host().ok_or_else(|| anyhow!("Invalid request"))?.to_owned();
       req.headers_mut().insert(header::HOST, HeaderValue::from_str(&org_host)?);
     };
@@ -127,7 +135,14 @@ where
     // apply upstream-specific headers given in upstream_option
     let headers = req.headers_mut();
     // apply upstream options to header, after X-Forwarded-For is added
-    apply_upstream_options_to_header(headers, &upstream_chosen.uri, upstream_candidates, &original_uri)?;
+    apply_upstream_options_to_header(
+      headers,
+      &original_uri,
+      original_host_header,
+      &upstream_chosen.uri,
+      upstream_candidates,
+      &self.globals.proxy_config.trusted_forwarded_proxies,
+    )?;
 
     // update uri in request
     ensure!(